Subindo com erro na hora de adicionar tarefas

+++ src/main/java/org/example/hydrocore/dto/request/AvisosRequestDTO.java
@@ -24,10 +24,10 @@ public class AvisosRequestDTO {
 
     @NonNull
     @Min(1)
-    private Long idEta;
+    private Integer idEta;
 
     @NonNull
     @Min(1)
-    private Long idPrioridade;
+    private Integer idPrioridade;
 
 }
